# openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerAccepted": {
        "properties": {
          "next_prompt_url": {
            "title": "Next Prompt Url",
            "type": "string"
          },
          "status": {
            "const": "processing",
            "default": "processing",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "next_prompt_url"
        ],
        "title": "AnswerAccepted",
        "type": "object"
      },
      "AnswerIn": {
        "additionalProperties": false,
        "properties": {
          "answer_text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Answer Text",
            "type": "string"
          },
          "client_submission_id": {
            "format": "uuid",
            "title": "Client Submission Id",
            "type": "string"
          },
          "turn_index": {
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Turn Index",
            "type": "integer"
          }
        },
        "required": [
          "turn_index",
          "answer_text",
          "client_submission_id"
        ],
        "title": "AnswerIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JoinIn": {
        "additionalProperties": false,
        "properties": {
          "join_code": {
            "maxLength": 12,
            "minLength": 1,
            "title": "Join Code",
            "type": "string"
          }
        },
        "required": [
          "join_code"
        ],
        "title": "JoinIn",
        "type": "object"
      },
      "JoinOut": {
        "properties": {
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "mission_title": {
            "title": "Mission Title",
            "type": "string"
          },
          "participant_id": {
            "format": "uuid",
            "title": "Participant Id",
            "type": "string"
          },
          "publication_id": {
            "format": "uuid",
            "title": "Publication Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "enum": [
              "lobby",
              "open"
            ],
            "title": "Run Status",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "warmup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WarmupOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "participant_id",
          "publication_id",
          "mission_title",
          "run_status",
          "session_id",
          "warmup",
          "deadline_at"
        ],
        "title": "JoinOut",
        "type": "object"
      },
...
      "MissionCreatedOut": {
        "properties": {
          "mission_id": {
            "format": "uuid",
            "title": "Mission Id",
            "type": "string"
          }
        },
        "required": [
          "mission_id"
        ],
        "title": "MissionCreatedOut",
        "type": "object"
      },
...
      "MissionIn": {
        "additionalProperties": false,
        "properties": {
          "knowledge_base_id": {
            "format": "uuid",
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "learning_objective": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Learning Objective",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "knowledge_base_id",
          "title",
          "learning_objective"
        ],
        "title": "MissionIn",
        "type": "object"
      },
...
      "ReflectionOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "mission_title": {
            "title": "Mission Title",
            "type": "string"
          },
          "opening_guess": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpeningGuessOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "mission_title",
          "completed_at",
          "content",
          "opening_guess"
        ],
        "title": "ReflectionOut",
        "type": "object"
      },
...
      "StateOut": {
        "properties": {
          "deadline_at": {
            "format": "date-time",
            "title": "Deadline At",
            "type": "string"
          },
          "probe_number": {
            "title": "Probe Number",
            "type": "integer"
          },
          "probe_total": {
            "title": "Probe Total",
            "type": "integer"
          },
          "prompt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflection_ready": {
            "title": "Reflection Ready",
            "type": "boolean"
          },
          "safety_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Message"
          },
          "server_now": {
            "format": "date-time",
            "title": "Server Now",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "turn_index": {
            "title": "Turn Index",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "turn_index",
          "probe_number",
          "probe_total",
          "started_at",
          "deadline_at",
          "prompt",
          "safety_message",
          "reflection_ready",
          "server_now"
        ],
        "title": "StateOut",
        "type": "object"
      },
      "StudentLobbyOut": {
        "properties": {
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "participant_status": {
            "title": "Participant Status",
            "type": "string"
          },
          "run_status": {
            "title": "Run Status",
            "type": "string"
          },
          "server_now": {
            "format": "date-time",
            "title": "Server Now",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "warmup_choice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Choice Id"
          }
        },
        "required": [
          "run_status",
          "participant_status",
          "warmup_choice_id",
          "session_id",
          "started_at",
          "deadline_at",
          "server_now"
        ],
        "title": "StudentLobbyOut",
        "type": "object"
      },
      "StudentMissionsOut": {
        "properties": {
          "completed": {
            "items": {
              "$ref": "#/components/schemas/MissionCardOut"
            },
            "title": "Completed",
            "type": "array"
          },
          "open": {
            "items": {
              "$ref": "#/components/schemas/MissionCardOut"
            },
            "title": "Open",
            "type": "array"
          },
          "upcoming": {
            "items": {
              "$ref": "#/components/schemas/MissionCardOut"
            },
            "title": "Upcoming",
            "type": "array"
          }
        },
        "required": [
          "upcoming",
          "open",
          "completed"
        ],
        "title": "StudentMissionsOut",
        "type": "object"
      },
...
      "StudentReflectionsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentReflectionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "StudentReflectionsOut",
        "type": "object"
      },
...
      "TelemetryAccepted": {
        "properties": {
          "accepted_client_seq": {
            "title": "Accepted Client Seq",
            "type": "integer"
          }
        },
        "required": [
          "accepted_client_seq"
        ],
        "title": "TelemetryAccepted",
        "type": "object"
      },
...
      "WarmupSavedOut": {
        "properties": {
          "choice_id": {
            "title": "Choice Id",
            "type": "string"
          },
          "scored": {
            "const": false,
            "default": false,
            "title": "Scored",
            "type": "boolean"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "choice_id",
          "submitted_at"
        ],
        "title": "WarmupSavedOut",
        "type": "object"
      },
      "WindowSessionOut": {
        "properties": {
          "deadline_at": {
            "format": "date-time",
            "title": "Deadline At",
            "type": "string"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptOut"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "started_at",
          "deadline_at",
          "prompt"
        ],
        "title": "WindowSessionOut",
        "type": "object"
      },
...
      "nalar__presentation__api__schemas__student__WarmupChoiceIn": {
        "additionalProperties": false,
        "properties": {
          "choice_id": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Choice Id",
            "type": "string"
          }
        },
        "required": [
          "choice_id"
        ],
        "title": "WarmupChoiceIn",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "__Host-nalar_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "NALAR Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/missions": {
      "post": {
        "operationId": "create_mission_api_v1_missions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Mission",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/v1/student/missions": {
      "get": {
        "operationId": "missions_api_v1_student_missions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentMissionsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Missions",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/publications/{publication_id}/window-session": {
      "post": {
        "operationId": "window_session_api_v1_student_publications__publication_id__window_session_post",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Publication Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Window Session",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/reflections": {
      "get": {
        "operationId": "reflections_api_v1_student_reflections_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentReflectionsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Reflections",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/runs/join": {
      "post": {
        "operationId": "join_api_v1_student_runs_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Join",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/runs/{run_id}/lobby": {
      "get": {
        "operationId": "lobby_api_v1_student_runs__run_id__lobby_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLobbyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Lobby",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/runs/{run_id}/warmup-choice": {
      "put": {
        "operationId": "warmup_choice_api_v1_student_runs__run_id__warmup_choice_put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/nalar__presentation__api__schemas__student__WarmupChoiceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarmupSavedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Warmup Choice",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/sessions/{session_id}/answers": {
      "post": {
        "operationId": "submit_answer_api_v1_student_sessions__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Submit Answer",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/sessions/{session_id}/reflection": {
      "get": {
        "operationId": "reflection_api_v1_student_sessions__session_id__reflection_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "description": "Evaluation still pending"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Reflection",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/sessions/{session_id}/state": {
      "get": {
        "operationId": "session_state_api_v1_student_sessions__session_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Session State",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/student/sessions/{session_id}/telemetry": {
      "post": {
        "operationId": "telemetry_api_v1_student_sessions__session_id__telemetry_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Telemetry",
        "tags": [
          "student"
        ]
      }
    },
...
  }
}